requirements.txt
cython
numpy
imagecorruptions
matplotlib
terminaltables
six
pycocotools
pillow==6.2.2
timm
